refactor: split admin settings and setup pages
