+++ packages/gpt2giga-ui/src/gpt2giga_ui/static/admin/pages/settings/types.js
@@ -1,0 +1,7 @@
+export const SETTINGS_LABELS = {
+    application: "Application",
+    observability: "Observability",
+    gigachat: "GigaChat",
+    security: "Security",
+    history: "History",
+};
