+++ packages/gpt2giga-ui/src/gpt2giga_ui/static/admin/pages/setup/types.js
@@ -1,0 +1,1 @@
+export {};
